fix:
## CountSizeOfWorkDir
### visitFile
>
&&!file.toFile().getName().equalsIgnoreCase("scan.tmp")
##
FileSystemWorker
### NO DEPRECATED
## NetScannerSvc
> Refactor

+++ modules/networker/src/main/java/ru/vachok/networker/fileworks/CountSizeOfWorkDir.java
@@ -57,7 +57,7 @@ public class CountSizeOfWorkDir extends SimpleFileVisitor<Path> implements Progr
 
 
     @Override public FileVisitResult visitFile(Path file , BasicFileAttributes attrs) throws IOException {
-        if(file.toFile().length() <= 0) {
+        if(file.toFile().length() <= 0 && !file.toFile().getName().equalsIgnoreCase("scan.tmp")) {
             try{
                 Files.deleteIfExists(file);
             }catch(Exception e){
